Add immediateAction row selection behavior option to payment sheet playground
