+++ Example/PaymentSheet Example/PaymentSheet Example/EmbeddedPlaygroundViewController.swift
@@ -37,6 +37,7 @@ class EmbeddedPlaygroundViewController: UIViewController {
     private let intentConfig: EmbeddedPaymentElement.IntentConfiguration
 
     private(set) var embeddedPaymentElement: EmbeddedPaymentElement?
+    private var paymentMethodsViewController: EmbeddedPaymentElementWrapperViewController?
 
     private lazy var loadingIndicator: UIActivityIndicatorView = {
         let indicator = UIActivityIndicatorView(style: .large)
@@ -84,6 +85,13 @@ class EmbeddedPlaygroundViewController: UIViewController {
         resetButton.translatesAutoresizingMaskIntoConstraints = false
         resetButton.addTarget(self, action: #selector(clearSelection), for: .touchUpInside)
         return resetButton
+    }()
+
+    private lazy var paymentMethodButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle("Select payment method", for: .normal)
+        button.addTarget(self, action: #selector(didTapPaymentMethodButton), for: .touchUpInside)
+        return button
     }()
 
     private let settingsViewContainer = UIStackView()
@@ -152,14 +160,25 @@ class EmbeddedPlaygroundViewController: UIViewController {
         scrollView.translatesAutoresizingMaskIntoConstraints = false
         view.addSubview(scrollView)
 
+        // If we are using the default row selection behavior, we include the payment element inline, otherwise we show a button to open show the payment element in a sheet
+        let paymentElementView = switch configuration.rowSelectionBehavior {
+        case .immediateAction:
+            paymentMethodButton
+        case .default:
+            embeddedPaymentElement.view
+        @unknown default:
+            fatalError("Implement how new row selection behavior should be displayed")
+        }
+
         // All our content is in a stack view
         let stackView = UIStackView(arrangedSubviews: [
             settingsViewContainer,
-            embeddedPaymentElement.view,
+            paymentElementView,
             paymentOptionView,
             checkoutButton,
             clearPaymentOptionButton,
         ])
+
         stackView.axis = .vertical
         stackView.translatesAutoresizingMaskIntoConstraints = false
         stackView.isLayoutMarginsRelativeArrangement = true
@@ -170,7 +189,6 @@ class EmbeddedPlaygroundViewController: UIViewController {
 #if DEBUG
         stackView.addArrangedSubview(testHeightChangeButton)
 #endif
-
         NSLayoutConstraint.activate([
             scrollView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             scrollView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
@@ -256,6 +274,22 @@ class EmbeddedPlaygroundViewController: UIViewController {
         embeddedPaymentElement?.clearPaymentOption()
     }
 
+    @objc
+    func didTapPaymentMethodButton() {
+        guard let embeddedPaymentElement else { return }
+        let paymentMethodsViewController = EmbeddedPaymentElementWrapperViewController(embeddedPaymentElement: embeddedPaymentElement, needsDismissal: { [weak self] in
+            self?.dismiss(animated: true)
+            self?.updatePaymentOptionView()
+        })
+        self.paymentMethodsViewController = paymentMethodsViewController
+        let navController = UINavigationController(rootViewController: paymentMethodsViewController)
+        present(navController, animated: true)
+    }
+
+    func updatePaymentOptionView() {
+        guard let embeddedPaymentElement else { return }
+        paymentOptionView.configure(with: embeddedPaymentElement.paymentOption, showMandate: configuration.embeddedViewDisplaysMandateText)
+    }
 }
 
 // MARK: - EmbeddedPaymentElementDelegate
